homelab-docker-network/pyproject.toml
[project]
name = "homelab-docker-network"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Vo Van Nghia", email = "git@vnghia.com" }]
dependencies = [
    "pulumi-docker>=4.6.1",
    "homelab-config",
    "homelab-docker",
]
requires-python = ">= 3.12"
classifiers = ["Private :: Do Not Upload"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/homelab_docker_network"]

[tool.mypy]
strict = true
